+++ extensions/fastmail-masked-email/.gitignore
@@ -3,5 +3,11 @@
 # dependencies
 /node_modules
 
+# Raycast specific files
+raycast-env.d.ts
+.raycast-swift-build
+.swiftpm
+compiled_raycast_swift
+
 # misc
 .DS_Store
