+++ extensions/fastmail-masked-email/src/create.tsx
@@ -1,42 +1,46 @@
 import {
-  showToast,
-  Clipboard,
-  Form,
   Action,
   ActionPanel,
-  Toast,
-  showHUD,
-  closeMainWindow,
+  Clipboard,
+  Form,
   Icon,
-  getPreferenceValues,
   PopToRootType,
+  Toast,
+  closeMainWindow,
+  getPreferenceValues,
+  showHUD,
+  showToast,
 } from "@raycast/api";
-import { createMaskedEmail } from "./utils";
+import { createMaskedEmail } from "./fastmail";
 
 type Preferences = {
   create_prefix: string;
 };
 
-type Values = {
+type FormValues = {
   prefix: string;
   description: string;
 };
 
-export default () => {
-  const preferences = getPreferenceValues<Preferences>();
-  const handleSubmit = async (values: Values) => {
+export default function Command() {
+  const { create_prefix } = getPreferenceValues<Preferences>();
+
+  const handleSubmit = async ({ prefix, description }: FormValues) => {
     const toast = await showToast({ style: Toast.Style.Animated, title: "Creating masked email..." });
+
     try {
-      const email = await createMaskedEmail(values.prefix, values.description);
+      const email = await createMaskedEmail(prefix, description);
+
       Clipboard.copy(email);
-      await toast.hide();
+      await showHUD("🎉 Masked email copied to clipboard");
       await closeMainWindow({ clearRootSearch: true, popToRootType: PopToRootType.Immediate });
-      await showHUD("🎉 Masked email address copied to clipboard");
-    } catch (e) {
-      if (e instanceof Error) {
-        await toast.hide();
-        showToast({ style: Toast.Style.Failure, title: "Error", message: e.message });
-      }
+    } catch (error) {
+      toast.style = Toast.Style.Failure;
+      toast.title = "Failed to create masked email";
+      toast.message =
+        error instanceof Error
+          ? error.message
+          : "An error occurred while creating the masked email, please try again later";
     }
   };
 
@@ -51,18 +55,18 @@ export default () => {
       <Form.TextField
         id="prefix"
         title="Prefix (Optional)"
-        placeholder="Prefix to use for this email address"
-        defaultValue={preferences.create_prefix}
-        info={`This field is optional. If you have configured a default prefix in the preferences, it will be used here. If you leave this field empty, no prefix will be used.
+        placeholder="Prefix to use for this masked email"
+        defaultValue={create_prefix}
+        info={`If you have configured a default prefix in the preferences, it will be used here. If you leave this field empty, no prefix will be used.
 
 A prefix must be <= 64 characters in length and only contain characters a-z, 0-9 and _ (underscore)`}
       />
       <Form.TextField
         id="description"
         title="Description (Optional)"
-        placeholder="What is this masked email address for?"
+        placeholder="What is this masked email for?"
         autoFocus={true}
       />
     </Form>
   );
-};
+}
